bugfix: fix DeepSeek V4 NPU DP graph execution.
