+++ xllm/core/layers/npu/npu_qwen3_moe_decoder_layer_impl.h
@@ -105,7 +105,8 @@ class NpuQwen3MoeDecoderLayerImpl : public BaseLayer {
                                torch::Tensor& attn_mask,
                                KVCache& kv_cache,
                                const ModelInputParams& input_params,
-                               bool is_prefill);
+                               bool is_prefill,
+                               bool use_graph_decode_input);
 
   torch::Tensor block_tables_placeholder_;
   std::string model_name_;
@@ -129,10 +130,12 @@ class NpuQwen3MoeDecoderLayerImpl : public BaseLayer {
 
   int32_t num_speculative_tokens_ = 0;
   atb_speed::qwen::MoeDecoderLayerParam prefill_param_;
-  atb_speed::qwen::MoeDecoderLayerParam decode_param_;
+  atb_speed::qwen::MoeDecoderLayerParam decode_graph_param_;
+  atb_speed::qwen::MoeDecoderLayerParam decode_eager_param_;
 
   atb_speed::Model::Node prefill_node_;
-  atb_speed::Model::Node decode_node_;
+  atb_speed::Model::Node decode_graph_node_;
+  atb_speed::Model::Node decode_eager_node_;
 
   atb::Tensor internal_tensor_;
