+++ xllm/core/runtime/forward_shared_memory_manager.cpp
@@ -2254,7 +2254,10 @@ inline void deserialize_forward_input_payload(
                 manager_table,
                 /*stream=*/nullptr,
                 /*force_host_materialize=*/true);
-    input_params.multi_block_tables.emplace_back(std::move(manager_table));
+    // Clone to decouple from shared memory buffer. With schedule_overlap the
+    // buffer may be overwritten by the next step while the worker is still
+    // reading multi_block_tables (which stay on CPU for DSA metadata).
+    input_params.multi_block_tables.emplace_back(manager_table.clone());
   }
 
   read_dit_forward_input(context, input_params.dit_forward_input);
